PUSH

+++ 42/cub3dMAGEULE/libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: chly-huc <chly-huc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/06 09:52:06 by chly-huc          #+#    #+#             */
-/*   Updated: 2020/01/12 01:40:13 by chly-huc         ###   ########.fr       */
+/*   Updated: 2020/11/06 19:50:23 by chly-huc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t		ft_strlen(const char *str)
+int		ft_strlen(const char *str)
 {
 	int		i;
